Some final nits including removing calling default as a script
